+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/UnusedAssignmentRule.java
@@ -175,6 +175,9 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
                 if (isIgnorablePrefixIncrement(entry.rhs)) {
                     continue;
                 }
+                if (isUsedLocalVarWithoutInitializer(entry, result.usedVariables)) {
+                    continue;
+                }
 
                 Set<AssignmentEntry> killers = result.killRecord.get(entry);
                 final String reason;
@@ -246,6 +249,11 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
             return !getProperty(CHECK_PREFIX_INCREMENT) && !(assignment.getParent() instanceof ASTStatementExpression);
         }
         return false;
+    }
+
+    private boolean isUsedLocalVarWithoutInitializer(AssignmentEntry entry,
+            Set<ASTVariableDeclaratorId> usedVariables) {
+        return entry.isDeclarationOnly() && usedVariables.contains(entry.var);
     }
 
     private static String makeMessage(AssignmentEntry assignment, /* Nullable */ String reason, boolean isField) {
@@ -1118,6 +1126,7 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
 
         final Set<AssignmentEntry> allAssignments;
         final Set<AssignmentEntry> usedAssignments;
+        final Set<ASTVariableDeclaratorId> usedVariables;
 
         // track which assignments kill which
         // assignment -> killers(assignment)
@@ -1129,9 +1138,11 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
 
         private GlobalAlgoState(Set<AssignmentEntry> allAssignments,
                                 Set<AssignmentEntry> usedAssignments,
+                                Set<ASTVariableDeclaratorId> usedVariables,
                                 Map<AssignmentEntry, Set<AssignmentEntry>> killRecord) {
             this.allAssignments = allAssignments;
             this.usedAssignments = usedAssignments;
+            this.usedVariables = usedVariables;
             this.killRecord = killRecord;
 
         }
@@ -1139,6 +1150,7 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
         private GlobalAlgoState() {
             this(new HashSet<AssignmentEntry>(),
                  new HashSet<AssignmentEntry>(),
+                 new HashSet<ASTVariableDeclaratorId>(),
                  new HashMap<AssignmentEntry, Set<AssignmentEntry>>());
         }
     }
@@ -1240,6 +1252,7 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
         }
 
         void use(ASTVariableDeclaratorId var) {
+            global.usedVariables.add(var);
             VarLocalInfo info = symtable.get(var);
             // may be null for implicit assignments, like method parameter
             if (info != null) {
@@ -1458,6 +1471,15 @@ public class UnusedAssignmentRule extends AbstractJavaRule {
         AssignmentEntry(ASTVariableDeclaratorId var, JavaNode rhs) {
             this.var = var;
             this.rhs = rhs;
+        }
+
+        public boolean isDeclarationOnly() {
+            if (var == rhs && var.isLocalVariable()) {
+                ASTVariableDeclarator declarator = var.getFirstParentOfType(ASTVariableDeclarator.class);
+                ASTLocalVariableDeclaration localVar = declarator.getFirstParentOfType(ASTLocalVariableDeclaration.class);
+                return !declarator.hasInitializer() && !(localVar.getParent() instanceof ASTForStatement);
+            }
+            return false;
         }
 
         @Override